+++ packages/std/src/results/init.rs
@@ -7,6 +7,67 @@ use crate::{Binary, Empty};
 use super::attribute::Attribute;
 use super::cosmos_msg::CosmosMsg;
 
+/// A response of the contract entry point `init`.
+///
+/// This type can be constructed directly at the end of the call. Alternatively a
+/// mutable response instance can be created early in the contract's logic and
+/// incrementally be updated.
+///
+/// ## Examples
+///
+/// Direct:
+///
+/// ```
+/// # use cosmwasm_std::{Binary, DepsMut, Env, MessageInfo, MigrateResponse};
+/// # type InitMsg = ();
+/// #
+/// use cosmwasm_std::{attr, InitResponse, StdResult};
+///
+/// pub fn init(
+///     deps: DepsMut,
+///     _env: Env,
+///     _info: MessageInfo,
+///     msg: InitMsg,
+/// ) -> StdResult<InitResponse> {
+///     // ...
+///
+///     Ok(InitResponse {
+///         messages: vec![],
+///         attributes: vec![attr("action", "init")],
+///         data: None,
+///     })
+/// }
+/// ```
+///
+/// Mutating:
+///
+/// ```
+/// # use cosmwasm_std::{coins, BankMsg, Binary, DepsMut, Env, HumanAddr, MessageInfo, MigrateResponse};
+/// # type InitMsg = ();
+/// # type MyError = ();
+/// #
+/// use cosmwasm_std::InitResponse;
+///
+/// pub fn init(
+///     deps: DepsMut,
+///     _env: Env,
+///     info: MessageInfo,
+///     msg: InitMsg,
+/// ) -> Result<InitResponse, MyError> {
+///     let mut response = InitResponse::new();
+///     // ...
+///     response.add_attribute("Let the", "hacking begin");
+///     // ...
+///     response.add_message(BankMsg::Send {
+///         to_address: HumanAddr::from("recipient"),
+///         amount: coins(128, "uint"),
+///     });
+///     response.add_attribute("foo", "bar");
+///     // ...
+///     response.set_data(Binary::from(b"the result data"));
+///     Ok(response)
+/// }
+/// ```
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct InitResponse<T = Empty>
 where
@@ -28,6 +89,30 @@ where
             attributes: vec![],
             data: None,
         }
+    }
+}
+
+impl<T> InitResponse<T>
+where
+    T: Clone + fmt::Debug + PartialEq + JsonSchema,
+{
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    pub fn add_attribute<K: Into<String>, V: Into<String>>(&mut self, key: K, value: V) {
+        self.attributes.push(Attribute {
+            key: key.into(),
+            value: value.into(),
+        });
+    }
+
+    pub fn add_message<U: Into<CosmosMsg<T>>>(&mut self, msg: U) {
+        self.messages.push(msg.into());
+    }
+
+    pub fn set_data<U: Into<Binary>>(&mut self, data: U) {
+        self.data = Some(data.into());
     }
 }