+++ packages/std/src/results/migrate.rs
@@ -8,6 +8,9 @@ use crate::types::Empty;
 use super::attribute::Attribute;
 use super::cosmos_msg::CosmosMsg;
 
+/// A response of the contract entry point `migrate`.
+///
+/// See [InitResponse][crate::InitResponse] for how this is used.
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct MigrateResponse<T = Empty>
 where
@@ -29,5 +32,29 @@ where
             attributes: vec![],
             data: None,
         }
+    }
+}
+
+impl<T> MigrateResponse<T>
+where
+    T: Clone + fmt::Debug + PartialEq + JsonSchema,
+{
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    pub fn add_attribute<K: Into<String>, V: Into<String>>(&mut self, key: K, value: V) {
+        self.attributes.push(Attribute {
+            key: key.into(),
+            value: value.into(),
+        });
+    }
+
+    pub fn add_message<U: Into<CosmosMsg<T>>>(&mut self, msg: U) {
+        self.messages.push(msg.into());
+    }
+
+    pub fn set_data<U: Into<Binary>>(&mut self, data: U) {
+        self.data = Some(data.into());
     }
 }