+++ packages/std/src/results/mod.rs
@@ -11,6 +11,7 @@ mod query;
 mod system_result;
 
 pub use attribute::{attr, Attribute};
+#[allow(deprecated)]
 pub use context::Context;
 pub use contract_result::ContractResult;
 pub use cosmos_msg::{wasm_execute, wasm_instantiate, BankMsg, CosmosMsg, StakingMsg, WasmMsg};