+++ CHANGELOG.md
@@ -27,6 +27,8 @@ and this project adheres to
 - cosmwasm-std: (requires `stargate`) Added new `CosmosMsg::Ibc(IbcMsg)` message
   type to use ibctransfer app or send raw ics packets (if contract has ibc entry
   points) ([#692], [#710])
+- cosmwasm-std: Add mutable helper methods to `InitResponse`, `MigrateResponse`
+  and `HandleResponse` which make `Context` obsolete.
 - contracts: added new `ibc-reflect` contract that receives channels and assigns
   each an account to redispatch. Similar idea to ICS27/Interchain Accounts (but
   different implementation) ([#692], [#711], [#714])
@@ -72,6 +74,11 @@ and this project adheres to
 [#697]: https://github.com/CosmWasm/cosmwasm/issues/697
 [#736]: https://github.com/CosmWasm/cosmwasm/pull/736
 [#690]: https://github.com/CosmWasm/cosmwasm/issues/690
+
+### Deprecated
+
+- cosmwasm-std: `Context` is deprecated in favour of the new mutable helpers in
+  `InitResponse`, `MigrateResponse` and `HandleResponse`.
 
 ## [0.13.2] - 2021-01-14